Clean up comparator functions
